requirements.txt
black
matplotlib
pre-commit
pypdf
pyrtlsdr
pyrtlsdrlib
